Updated and added new pCT user functions
Finished (hopefully) the bash functions needed to manage code/data and
perform the various reconstruction tasks on Tardis.  Updated
instructions/documentation is soon to come.

The next step is to combine these (when possible/appropriate) and
generate tools for complete tasks which can then be used to copy
code/data to/from Tardis, run reconstruction (soon preprocessing too),
and ultimately share their data with as few tools and as little user
input as possible.