Merge data macro simplified
Implemented and tested rewrite of code merging RSP CSV data using the functions written for the TV table construction.  This confirms they were written appropriately generic to work for the very different ROI data input, which have different #s of elements, data organization, title/column labels, and different #s of tables in the horizontal and vertical directions.  This did require the merged RSP data array to be resorted, but this same sorting and function usage applies directly to the RSP error and standard deviation data as well, so this simplification can be applied to these by simplyi copying the RSP merge block of code and replacing the appropriate variable names.