Merge data macro: simplified usage + multiplot merge finished
(1) Finished replacing explicit results table construction code for RSP Error and Standard Deviations CSV data with the associated functions written as an encapsulation of the original explicit code, separated into individual functions for each task, e.g., adding empty rows/columns, titles & column labels, and rows/column/tables of data.

(2) Finished the code to merge the newly created merged CSV data files for each multiplot parameter value into single files.  The merged CSV data files apparently cannot be read from disk to the results table,  presumably due to the mixing of int, float, string, and empty elements, so the merged CSV data files are also written to TXT so they can be read back in for each multiplot parameter value and used to fill and subsequently export the results table.

Merging of the ROI analysis data for each multiplot parameter value yields a grid of tables: the rows of the grid correspond to multiplot parameter values and the columns of the grid correspond to the ROIs.  This is convenient as all the data for a particular ROI and iteration are in the same column.  The initial merged CSV data files have ROI tables arranged vertically instead so more of the data can be seen at once, but this offers an alternative arrangement that may be beneficial at times and the data can be accessed from the multiplot merged CSV file when it's not.