CSV data merge macro: encapsulation of multiplot data merge code as generic function
The code used to construct the grid of CSV data tables in the results table for each ROI analysis measure follows the same basic series of steps so this process was encapsulated and made into a generic function.  This function is now used to merge CSV data from each multiplot parameter value for each ROI analysis measure, reducing the procedure down to 2 function calls: 1 adds the title lines and the other adds the grid of tables, with the results table and corresponding array data written to CSV & TXT respectively.