Identified error, repeated merge_data changes, and added table grid insertion function
The issue with ImageJ was due to a Windows 10 update consuming all available system memory, so when files were attempted to be written to disk, it could only write the avialable number of bytes to the file and when this was 0 the resulting file was completely empty.  All changes to the ImageJ macros were lost as a result, but this was fortunately not a huge amount since commits have been made fairly regularly lately.

This commit now adds a more generic function for creating a grid of tables.  The # of tables per grid row is specified and based on the # of rows/columns per table and the total data size, the total number of tables and the corresponding # of table grid rows is determined.  Beginning with this new function and to be applied to others, the way in which the data is ordered in the data array from which the grid of tables is created is passed as a parameter and used to extract the appropriate data as each row of the tables is constructed.  This can either be in row order (equivalent to table order) or it can be in column order (where each row of the various tables are concatenated).

This commit also includes some slight modifications to the functions that add empty rows, the column summation rows, and the title/column header rows to account for table grid dimensions and for different column/row/title/header separation choices (as some of these only worked for the default single row/column separations).

Once the other functions are changed similarly, fewer functions will be needed to construct a grid of tables with any dimensions and both the 1st and 2nd level of merging should be able to be done with the same functions by appropriately concatenating and ordering the data, making it easier to add new or change existing merging schemes.  One new merging scheme that will be added is placing all ROI analysis measure data in a single file so the RSP, RSP error, std dev, and TV data can all be found in the same file and compared simultaneously.