merge_data: added new/fill/init_array functions w/ L = 1 support and simplified table_grid_{add_row,fill} functions
Generic array functions added to create an empty array of length L, fill an array with some element, and a combo of these with an init_array function.  The default Array.<x> functions do not work properly for an array length L = 1, so these new functions add support for this.

The 2 table_grid_<x> functions were simplified to make subsequent usage easier.  The table_grid_add_row function now makes use of a new table_add_row function which combines the addition of the row label, the row data added with table_fill_columns, and the skipping of columns between tables.  2 Booleans arguments at the end also control the column skipping before the first table on each table grid row and the suppression of column skipping after the last table on a table grid row.  A conditional statement within parentheses is passed directly to the function call rather than evaluating the conditional and assigning it to a variable since there is no other use for these Booleans and therefore no point in storing their result.

The merging of the CSV data files is now being performed with the table_add_title_lines and table_grid_fill functions, thereby reducing the number of lines of code and making the logic for these easier to follow.  The next step is to replace the usage of the table_multiplot_fill function with the table_grid_fill and related functions to further simplify the code and logic.