merge_data.ijm: Replaced table_multiplot_fill with table_grid_fill for ROI multiplots
The table_grid_fill function is now used for every data merging process except the TV multiplot, which does not need any additional function additions or modifications, it simply needs to be implemented.  The multiplot data merging only required 1 slight modification of the grid fill function: since the same tables are created from multiple files, the table titles are repeated for each multiplot value.  To account for this, the grid fill function now checks to see if the # of tables exceeds the size of the titles array and duplicates these as many times as necessary to make them equal.  Additionally, outside the function and prior to the multiplot merging, an array is created with the required # of duplications of the titles and this is passed to the function instead.  The function will duplicate these itself but since this same process needs to be repeated for each ROI analysis, it is less computational work to simply construct this and pass it to the function rather than have the function create it each time and the resulting code is easier to read this way as well.

Once this has been implemented for the TV multiplot, all unnecessary functions will be removed and the other data merging processes will be added, such as merging all ROI analysis and TV data into a single file from which all data for a particular reconstruction and/or multiplot reconstructions can be viewed simultaneously.