merge_data.ijm: further simplified code, reduced lines of code, and cleaned up code
Created new function table_grid_fill_all which is in the form that receives the title/table dimensions and configuration variables encapsulated within arrays constructed by combining variables into a single array and then extracting the indifvidual variable values from within the functions.  This function adds the title lines above the grid of tables then fills the table grid, unlike the table_grid_fill function which requires the grid titles be added separately before it is called.

Another similar new function is table_grid_fill_all_imported which has the same functionality as the table_grid_fill_all function, but the list of input data file paths from which the data is imported is passed to the function instead of the data itself.  The data array is then constructed by importing the data from each of the files in the list of file paths and the remainder of the function then proceeds in the same way as table_grid_fill_all.

A new function table_add_colsumn_row_label was also created which determine the length of the longest row label string of each table in a table grid and this information is then passed to the table_add_colsum_row/rows functions and used to add a series of dashes and a right carat to the COLUMN_SUM_ROW_LABEL string such that the resulting label of the column sum row include the string "Sum " followed by a right arrow that extends to the edge of the results table column.  This helps make this row stand out from the others and be immediately distinguishable from the raw table data.  However, since the dash character is smaller than most other characters, 1.3 dashes are added for every regular character that we wish to extend the label by, though this factor is only true on average and a particular label may contain characters that are larger or smaller than average and the resulting arrow may extend slightly more or less than desired, but the difference is small enough that in no case thus far encountered has this made a noticeable difference.

New CSV/TXT data file writing functions were also created which accept the Boolean controlling whether the file write is to be performed as an argument rather than requring the corresponding write statements to be placed inside an if statement block.

Finally, the variables associated with the ROI/TV data and multiplot data merging tasks which remain constant throughout execution were moved to the upper portion of the program with the other variable definition, reducing the # of lines of code in the data merging task portion of the program .  In addition, I scanned through the program looking for opportunities to clean up the code and remove unnecessary/commented lines of code.