merge_data.ijm: modified data merge folders/filenames and added dir creation
The multiplot data merging files were not being written to disk because their destination folders did not exist.  In addition, many of the output file names/folders were not set as desired, so these were modified and their corresponding variables names changed to a consistent and clear naming scheme.  It was also discovered that the table grid titles set during the CSV data merging were not correct, so these were modified appropriately as well.  The multiplot data merging results are also now written to a 2nd output directory named with the current parameter values and the parameter value ranges in the form X[#-#]_ are appended to the folder name at the string location corresponding to this parameter (these range parameters are Nk and A/TV/L for these 3 multiplot parameter value tests).

Other variables were given new names and more that are not in use were identified and removed in this commit.  Several ImageJ "run" commands were also encapsulated within new function definitions such that they accept function arguments and use these to construct the string argument of the "run" command rather than continuing to have to write the string explicitly each time.  Lastly, a few more functions were written whose function and action depend on whether the macro is executed from within the macro or if the macro is externally executed with the runMacro command from within another macro.  The way windows are closed at the beginning of execution, the way execution is halted after a user defined number of loop executions, and the way the multiplot_parameter_prefix is set varies depending on whether it is an internal/external macro execution: external macro executions pass multiplot_parameter_prefix  as a macro execution argument so the corresponding variable is set by reading the input macro argument, thereby allowing the merge_data_control macro to iteratively call merge_data with each of the multiplot parameter prefixes such that all possible data merging is generated with a single execution of merge_data_control.

A few small modifications were also made to the other 2 primary analysis macros, but none of these were at all consequential.