ImageJ parameter value string construction/manipulation
There are numerous functions for constructing filenames and strings from parameter values as well as many others that extract parameter values from directory and other strings.  Each of these are specialized for a particular scenario but in general are composed of roughly the same code.  Hence, the objective has been to create a more generic function that can perform as many of the different string constructions as possible by adding additional function arguments and corresponding conditional control within the function.  The variations that are now supported are

(1) specifying the character separating the parameter prefix and value/range and the character separating each parameter prefix+value term ('_', '=', '-', ',', etc.)
(2) specifying if parameters with a single value are to be included or not, which parameters should be excluded such as the multiplot or target test parameters ("L"/"FBP_r"/"TV"/"A") & "Nk" for NTVS, respectively)
(3) specifying which parameters should be inserted with their parameter value range instead of a single value ("[<prefix>#-#]" for multivalue parameters; "<prefix>[#]", "<prefix>=#", or omitted for single value parameters)
(4) specifying if trailing 0s on a parameter value string should be removed or not

The function has been tested for some applications/configurations but it has only been used to replace 1 string construction function so it will need to be tested further as itt is used to iteratively replace other functions.  A few helper functions were also created and a few other existing functions were modified to provide the support needed by the functions.  For example, functions were created to remove particular indices of an array (used to remove the multiplot parameter and/or target test parametter index from an array of parameter value indices so the resulting array can be used inside the function to control which parameters are omitted or included as a parameter value range).

It probably won't be possible to replace all existing functions with this new function, but once these other functions have been replaced, the generalization process will be repeated for those remaining functions to try to replace as many of these wiith a single function as well.  This entire process will then be repeated as many times as necessary to minimize the # of required functions and replace as many specialized functions with generic flexible functions as possible.