merge_data.ijm: Created more generic functions to replace specialized functions
New functions were created to prepend/append the same strings to each element of an array.  A few more array_<X> functions were also created to help with this and other tasks:
(1) a function that expands the length of an array, or creates an array if the input data is a single element, to match the length of another array, filling the new portion of the array with the last element of the initial array or the single value when the input is a single element instead of an array.
(2) a function that applies a mathematical operation to each value of an array with a single common value, supporting addition, subtraction, multiplication, and division and the ability to switch the order of the operands since for subtraction/division, the result depends on whether the array element or the common value is used as the left/right operand.
(3) a function to determine if 2 arrays are equal, checking each element 1 by 1.
(4) a function to determine if 2 strings are equal

The primary new function was thoroughly tested and validated using the array_match function to verify that the arrays resulting from the initial code implementation are the same as those generated using the new generic function.  The initial implementation will now be removed and replaced with this new implementation for the tasks performed before the data merging loop begins since this is working properly.

The new parameter value string/directory construction/manipulation functions have been validated for the 1st loop iteration but it is important to validate these for each loop iteration and throughout the entire execution since any small mistake can have drammatic consequences; this could overwrite existing data with invalid data and/or create new improperly named/located files that will be time consuming to identify and remove later.  Hence, before permanently replacing the various specialized functions used inside the loop (to construct file names/paths to read from and write to and the strings used for labeling) with the generic functions designed to provide the same functionality given the appropriate function arguments, these will need to be validated for the entirety of execution using the string/array match function to identify any instances where the new functions do not produce the same results as those in the initial implementation.  Once these have been completely validated, the old implementation will be completely removed and any functions are no longer used anywhere in the program will also be deleted.  Ideally all of the specialized functions will be able to be replaced but this may require additional generic functions and it's possible that there are some specialized functions that have no easy replacement scheme.  In any case, thus far every specialized function for constructing/manipulating file name/path and parameter value strings has been able to be replaced by 1 of the few generic functions created and the remaining specialized functions appear to be replaceable with functions at least similar to the new generic functions, so the bulk of the procedures for performing the corresponding string tasks have already been worked out and implemented and only slight modifications or expansions of these will be needed to replace those specialized functions that remain and themselves are actually replaceable.