merge_data.ijm: Validated generic function replacements of specialized functions
Added calls of string/array match functions to test each use of the new generic functions and compare the resulting string/array with those produced by the original implementation with the specialized function and allowed the program to run in entirety, with all data merging and file writing suppressed to prevent any invalid data writes.  The string/array matching functions have a Boolean argument which provides the ability to notify the user of a mismatched srting/array through a dialog window, allowing the user to continue execution by selecting "OK" or halting the macro by selecting "Cancel".  Hence, if execution finished without a dialog window opening, then there were no mismatched string/arrays.

Since no dialogs were opened and the new generic functions produced the same results as the specialized functions did for the entire execution, these new functions function as intended and the original specialized functions and corresponding statements can now be removed completely, along with the definitions of any functions that will no longer be used anywhere once the other functions have been removed.

None of the functions or the string/array matching and priting functions have been removed in this commit, that way the status of the program at the time of validations and the corresponding validation code can be saved and retained for the future in case an undiagnosed error went unidentified or for reference when attempting to validate other developments.