merge_data.ijm: Identified and corrected invalid string generated
The strings used to set the table grid titles in the multiplot data merging were being set incorrectly, with the mulitplot parameter value remaining constant during the merging for each multiplot parameter value.  This was due to both the use of a single folder input for parameter value extraction and (2) usage of the wrong parameter value array for constructing the combined parameter value string from it.

This was missed previously because this was one of only 2 usages of the new functions appearing inside the conditional blocks of code performing the data input, merging, and output, since these blocks were not executed during validation.  To validate these instances as well, the corresponding blocks of code were allowed to run and the writing/overwriting control Booleans were set to prevent and disk writing that may produce invalid data files if there are additional unidentified errors.

It now appears that all instances have been validated.  Also included in this commit is a modification of the multiplot data merging which had a single ROI analysis grid title array and therefore was adding this same set of strings to the results table, but these should and now are specific to the ROI analysis type and indicate the TXT data files from which the data for each multiplot parameter value were read from.  The next step is to remove any of the specialized functions no longer in use from the macro, rename the new generic functions appropriately in cases where the current name is simply the old name with the # 2 appended to the end, and begin identifying any other specialized functions that can now be replaced as well and which others can be merged into a new generic function.  The status of the macro after validating remains in this commit, like last commit this is to allow reverting back to this validation state if more errors should arise, but the first step beyond this commit is to remove all of the validation functions and corresponding print statements.