ROI_Analysis.ijm: Updated from Baylor_ICTHUS repo + merge_data.ijm: Removed unused specialized string functions & minor cleaning/simplifications
ROI_Analysis.ijm:

The version of the ROI_Analysis macro that has been being used is the one from the Baylor_ICTHUS/pCT_Reconstruction GitHub repository although every other macro being developed further and used are from the general and more appropriate collaboration wide pCT_Collaboration/pCT_Tools repository.  The pCT_Reconstruction repo is in a fair amount of disarray in terms of the numerous branches being out of sync with commits and needing to be merged, in fact the Matlab/ImageJ macros are not present in every branch now, so given this and the fact that tools such as the Matlab scripts and ImageJ macros are more appropriately maintained in the pCT_Tools repo accessible by the entire pCT collaboration.

Hence, the outdated version of the ROI_Analysis macro in pCT_Tools was updated to the state of this macro in the pCT_Reconstruction repo and future development will only be to the macros in the pCT_Tools repo.  At some point, ideally in the near future, all Matlab/ImageJ/other tools in the pCT_Tools repo will be updated/copied from those in the pCT_Reconstruction repo (as needed and where appropriate) and then all tools will be removed from pCT_Reconstruction such that all tools are exclusively available from pCT_Tools and nowhere else and pCT_Reconstruction contains only those files relevant to the reconstruction program.

merge_data.ijm:

The previous commit left everything from the process of verifying that the new generic parameter value string construction/manipulation functions and this has all been removed now, including the specialized function definitions that are no longer used.  Other modifications include cleaning/simplifying the code such as renaming some variables to more clearly identify their content, moving (hopefully at least most of) the single/double line definitions of very basic system/string/array/printing functions next to each other and placing this group above all of the other function definitions immediately below the end of the macro definition (instead of the existing alphabetic ordering by function name).  These functions are often referred to and this is much easier when they aren't buried among the other much larger function definitions and are instead all grouped together at an easy to scroll to location.

A few new functions were also defined or added from ROI_Analysis, most of which are single/double line basic functions, but the process by which the tab string is defined using the '\t' escape sequence or explicitly as a series of spaces was encapsulated as a function and the array_add_2_strings function was modified such that the strings to prepend/append can now also be an array of strings as well as a single common string for each.  To make it possible to prepend/append the strings using a single procedure for both the single common strings and string array input arguments, the single common strings are duplicated to build an array with the same length as the input string array so the strings being prepended/appended are always indexed from an array and separate procedures for these 2 possibilities is not necessary.  This also means that if the strings to prepend/append are passed as string arrays but the length of these is not the same as the input string array, they will be expanded to have equal lengths and the expanded portions of these arrays will be filled with the last element of the input arrays.  This is currently the only option to fill the new elements of the arrays, but the goal is to provide the ability to fill the new positions cyclically from the initial input array, effectively duplicating the input arrays as needed such that the arrays all have the same length.  Explicitly duplicating the arrays will not work in all cases though, since this only yields an array with the desired length if the initial and desired lengths are integer multiples of each other.  When they are, filling the new portion of the array with the cyclically acquired term of the initial input array yields the same result as duplication, but this procedure also works properly when directly duplicating the input array will not produce the desired array length. However, it may be possible to duplicate the input array such that the resulting array is as large or larger than the desired length if the loop over the length of the input string array will remain unaffected by the size of the prepend/append string arrays , then only the portions of these arrays up to the length of the input string array will be accessed during the loop and since the portion of these arrays beyond the length of the input string array will never be accessed, so the fact that they are longer than needed will not matter.

Adding another argument to control how the expanded portions of the arrays are filled is not ideal, there are already several arguments, so some way to specify this without an additional argument will be attempted.  If this cannot easily be implemented otherwise, the current duplication of the last element may simply be replaced with the full array duplication since this seems to be the more likely choice; I can think of scenarios where the duplication of every element would be desired but I cannot think of a scnerio when one would want to simply duplicate the last element.

In any case, the approach for expanding the length of the arrays to match each other may be delayed beyond the next commit since the reason array_add_2_strings was modified to also accept prepend/append string arrays was so it could also be used to replace the function that produces strings combining the parameter prefix, equals sign, and parameter value (typically used to set variables with "<...>parameter_equsls_strings<...>" in the variable name) and this requires separate string arrays for both the parameter prefixes and parameter values.  Development on this will now continue but since the input prepend/append strings will only be either single strings or string arrays with the desired length already, so the process by which the arrays are expanded to the length of the input string array is inconsequential and if it were implemented, this is just another portion of the function that would also need to be changed as the function is modified/expanded to iteratively replace more of the specialized functions currently still being used one at a time.  Expanding the new generic function capabilities and replacing the specialized functions is the focus right now and unless the array expansion via duplication becomes an integral part of this, implementation of any aspect of this will be delayed until a final version of the function is done and no major modifications of it that would require this procedure to be modified as well are likely anymore.