Added cp/rsync command aliases to back_profile scripts and continued ImageH macro development/standardization
Added command aliases for copying data with/without overwriting with cp/rsync to the load_pct_functions.sh script loaded by .bash_profile and modified the file/directory count functions fcount/dcount.

Continued development of the ROI_analysis_control.ijm macro and added a new macro development_testbench.ijm that is used for testing new code.  This is meant to contain all variables and functions available in one or more of the other macros.  Variables corresponding to constant values are intended to have capitalized names, but this has only been implemented in a couple of the macros, so this macro contains both the capitalized and lowercase versions of these variable definitions.  The hope is to iteratively replace the lowercase variable names with their capitalized versions in all macros so all macros use the same variable/function naming and definitions such that code can easily be added from one macro to another without other changes being needed.  The hope is to use the new testbench macro to aid in these efforts.