Began separating constant variable definitions from load_pct_functions.sh into a separate script
The .bash_profile related scripts are being restructured such that the definitions of variables that remain constant and any other similarly consistent definitions are sourced from a separate script, that way access to variables like these users should not change can be restricted so they can't and when these are removed from the other scripts, these are much shorter, cleaner, and easier to view/understand/modify.

Further separations may be desired in the future.