Improved ImageJ analysis tools and updated pCT user .bash_profile sourced scripts
ImageJ:

Finished rewriting the conditionally controlled ROI analysis routines which determine the range of reconstructions on which analysis if performed depending on the user controlled set of Booleans.

merge_data is also in the process of being expanded to incorporate secondary data which the consolidated data is to be compared with. At the moment, the data from multiple varying parameters can be consolidated and written to disk at increasingly larger levels, 1st every ROI/iteration for each value of the primary test parameter (NTVS: Nk) down to single files, then merging these together for each value of another parameter into single files, then merging these ALL together into separate single files for each quality measure.  Eventually the merging performed in this last step may be broken into individual steps for merging with each additional parameter with varying values, and come up with a standardized way for identifying the files to merge and naming the merged files such that this process is essentially an iterative process performing the same merging process at each level and simply repeating this for as many varying parameters as there exists until the output is a single file, only differing in the order in which the parameters are merged together if the user desires and specifies a particular order to use. At the moment, these merging processes are fairly different especially the 1st as this requires significant restructuring of the data read from the ROI analysis files and direct construction of the tables in the "Results" window before saving to CSV, but subsequent merging from there would be more feasibly generalized into a standard routine without outrageous work as long as the file naming process was abstracted out into specific functions designed to be called at each level of merging. For now, the simple fact that merging works is sufficient.

ImageJ also now provides a config file for specifying the reconstruction info and desired analysis settings in the same way phantom ROIs are defined: users now specify the simulated/experimental, the dimensions of the image. number of iterations, and the range of iterations/slices to analyze and whether to flip the image horizontally/vertically in analysis.cfg. Hence, any hardcoded information that often changes for different reconstructions are now set using 1 of the config files so users need not search through the macros to determine and find the variables that need to be changed.

.bash_profile sourced scripts

The constant definition variables, i.e. those that remain fixed such as permanent file/directory and git repository/branch names and paths, node info, bash colors/commands, etc have been separated from the alias/function definitions and into their own file. pct_user_script is also being adapted to allow for arbitrary prompt/session appearance settings and even selection from a set of predefined settings while still performing all of the same tasks/definitions. The predefined prompt/appearance settings options will be added in the next commit, this commit is only intended to save the repo status after having verified proper separation of variables into a new script before the commented out overlap and other less important tasks are performed.